refactor : validation 적용

+++ store-api/src/main/java/com/zerobase/storeapi/controller/CartController.java
@@ -8,8 +8,14 @@ import com.zerobase.storeapi.domain.redis.form.UpdateOptionCartForm;
 import com.zerobase.storeapi.service.CartOrderService;
 import com.zerobase.storeapi.service.CartService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/store/cart")
@@ -21,30 +27,48 @@ public class CartController {
 
     @PostMapping
     public ResponseEntity<?> addCart(@RequestHeader(name = "Authorization") String token,
-                                     @RequestBody AddItemCartForm form){
+                                     @RequestBody @Validated AddItemCartForm form, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            List<String> errors = bindingResult.getAllErrors().stream().map(DefaultMessageSourceResolvable::getDefaultMessage).collect(Collectors.toList());
+            return ResponseEntity.badRequest().body(new ErrorResponse("400", "Validation failure", errors));
+        }
         return ResponseEntity.ok(cartService.addCart(memberClient.getMemberId(token), form));
     }
 
     @GetMapping
-    public ResponseEntity<?> getCart(@RequestHeader(name = "Authorization") String token){
+    public ResponseEntity<?> getCart(@RequestHeader(name = "Authorization") String token) {
         return ResponseEntity.ok(cartService.getCart(memberClient.getMemberId(token)));
     }
 
     @PatchMapping
     public ResponseEntity<?> deleteCartOption(@RequestHeader(name = "Authorization") String token,
-                                              @RequestBody DeleteOptionCartForm form) {
+                                              @RequestBody @Validated DeleteOptionCartForm form
+            , BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            List<String> errors = bindingResult.getAllErrors().stream().map(DefaultMessageSourceResolvable::getDefaultMessage).collect(Collectors.toList());
+            return ResponseEntity.badRequest().body(new ErrorResponse("400", "Validation failure", errors));
+        }
         return ResponseEntity.ok(cartService.deleteCartOption(memberClient.getMemberId(token), form));
     }
 
     @PutMapping
     public ResponseEntity<?> updateCartOption(@RequestHeader(name = "Authorization") String token,
-                                              @RequestBody UpdateOptionCartForm form) {
+                                              @RequestBody @Validated UpdateOptionCartForm form
+            , BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            List<String> errors = bindingResult.getAllErrors().stream().map(DefaultMessageSourceResolvable::getDefaultMessage).collect(Collectors.toList());
+            return ResponseEntity.badRequest().body(new ErrorResponse("400", "Validation failure", errors));
+        }
         return ResponseEntity.ok(cartService.updateCartOption(memberClient.getMemberId(token), form));
     }
 
     @PostMapping("/order")
     public ResponseEntity<?> orderCart(@RequestHeader(name = "Authorization") String token,
-                                       @RequestBody Cart cart) {
+                                       @RequestBody @Validated Cart cart, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            List<String> errors = bindingResult.getAllErrors().stream().map(DefaultMessageSourceResolvable::getDefaultMessage).collect(Collectors.toList());
+            return ResponseEntity.badRequest().body(new ErrorResponse("400", "Validation failure", errors));
+        }
         return ResponseEntity.ok(cartOrderService.orderCart(token, memberClient.getMemberId(token), cart));
     }
 }
